Sujet II/requirements.txt
numpy
pandas
scikit-learn
joblib
pyyaml
